+++ shared/core/src/lib/utils/deprecated.ts
@@ -1,7 +1,5 @@
-import * as lodash from 'lodash-es';
-
 import moment from 'moment';
 import 'moment/dist/locale/zh-cn';
 moment.locale('zh-cn');
 
-export { lodash, moment };
+export { moment };