+++ shared/framework-kernel/src/lib/stores/index.ts
@@ -1,3 +1,5 @@
 export * from './settings';
 export * from './router';
 export * from './tab-view';
+export * from './authentication';
+export * from './crypto';