+++ shared/bpmn-designer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @herodotus-cloud/bpmn-designer
 
+## 1.7.1
+
+### Patch Changes
+
+- 调整认证安全相关内容
+- Updated dependencies
+  - @herodotus-cloud/components@1.7.1
+  - @herodotus-cloud/bpmn-apis@1.7.1
+  - @herodotus-cloud/form-apis@1.7.1
+  - @herodotus-cloud/core@1.7.1
+
 ## 1.7.0
 
 ### Minor Changes