+++ shared/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @herodotus-cloud/core
 
+## 1.7.1
+
+### Patch Changes
+
+- 调整认证安全相关内容
+
 ## 1.7.0
 
 ### Minor Changes